Updated and improved convergence plot tool
