+++ modules/level_set/test/tests/verification/1d_level_set_mms/convergence.py
@@ -7,26 +7,5 @@
 #*
 #* Licensed under LGPL 2.1, please see LICENSE for details
 #* https://www.gnu.org/licenses/lgpl-2.1.html
-
-import sys, os
-import glob
-import pandas
-import numpy as np
-from mooseutils.ConvergencePlot import ConvergencePlot
-
-# The csv files to read
-filenames = glob.glob('level_set_mms*.csv')
-
-# Extract the data
-n = len(filenames)
-error = np.zeros(n)
-length = np.zeros(n)
-for i, filename in enumerate(filenames):
-    csv = pandas.read_csv(filename)
-    error[i] = csv['error'].iloc[-1]
-    length[i] = csv['h'].iloc[-1]
-
-
-fig = ConvergencePlot(length, error, xlabel='Element Length', ylabel='L2 Error')
-fig.save('convergence.pdf')
-fig.show()
+import mms
+mms.plot('level_set_mms*.csv', xlabel='Element Length', ylabel='L2 Error', output='convergence.pdf')
