+++ python/mms/__init__.py
@@ -8,3 +8,4 @@ else:
     from fparser import FParserPrinter, fparser, print_fparser
     from moosefunction import MooseFunctionPrinter, moosefunction, print_moose
     from evaluate import evaluate
+    from ConvergencePlot import ConvergencePlot, plot
